requirements.txt
# AviationStack API Crawler Dependencies

# HTTP requests
requests>=2.28.0

# YAML config parsing
pyyaml>=6.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For data analysis
pandas>=1.5.0